% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_kliep.R
\name{dr_kliep}
\alias{dr_kliep}
\title{Density-ratio SDM estimation with KLIEP}
\usage{
dr_kliep(
  presence_data = NULL,
  background_data = NULL,
  projection_data = NULL,
  sigma = "auto",
  kernel_num = 100,
  fold = 5,
  verbose = FALSE,
  method,
  object = NULL
)
}
\arguments{
\item{presence_data}{dataframe of covariates}

\item{background_data}{dataframe of covariates}

\item{projection_data}{dataframe of covariates}

\item{sigma}{Sigma parameter for KLIEP. Default is the KLIEP default.}

\item{kernel_num}{kernel_number for KLIEP. Default is the KLIEP default.}

\item{fold}{Number of folds for KLIEP cross-validation. Default is the KLIEP default.}

\item{method}{one of either "fit" or "predict"}

\item{object}{fitted object returned by a dr_... function. Only needed when method = "predict"}
}
\description{
dr_kliep is an internal function for density-ratio estimation with KLIEP.
}
\keyword{internal}
